Read encrypted config during startup
When the service starts up, spawn the http server to retrieve a
passphrase. It is also possible to start the server without an encrypted
configuration file by providing the `--disable-encryption` flag. If
encryption is used, the config file has to be located in
`config.yaml.enc`, and if encryption is disabled, the config file will
be loaded from `config.yaml`.