Display author information
To make it easier to tell which deployment is currently queued up, the
name of the author is added to the display. To display custom names, the
authors section of the config can be populated with `fullName` and
`alias` entries. If no author alias is found, the first name of the
author string is used (since restricted to 16 characters)